+++ src/ClampMotor.h
@@ -10,7 +10,7 @@ public:
     void setEffort(int effort);
     void setEffortWithDeadband(int effort);
     byte moveTo(int position);
-    long getPosition();
+    int getPosition();
     void setup();
     void reset();
